refactor: extract cookie consent logic into reusable mixin
